requirements.txt
akeyless==3.2.3
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decouple==3.5
fastapi==0.75.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
hvac==1.0.2
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
logtail-python==0.1.3
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
py==1.11.0
pydantic==1.9.0
pyhcl==0.4.4
pyparsing==3.0.7
pytest==7.1.0
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2021.3
requests==2.27.1
rollbar==0.16.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
